finalização user

+++ MyPrecious/src/dao/impl_DB/MovimentacaoDaoBD.java
@@ -1,10 +1,9 @@
 package dao.impl_DB;
 
+import controller.LoginController;
 import dao.CategoriaDAO;
 import dao.MovimentacaoDAO;
 import dominio.Categoria;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.time.LocalDate;
@@ -31,8 +30,8 @@ public class MovimentacaoDaoBD extends DaoBd<Movimentacao> implements Movimentac
     public void salvar(Movimentacao mov) {
         try {
             int id = 0;
-            String sql = "INSERT INTO movimentacao (valor, data, hora, descricao, tipo, idcategoria) "
-                        + "VALUES (?,?,?,?,?,?)";
+            String sql = "INSERT INTO movimentacao (valor, data, hora, descricao, tipo, idcategoria, user_id) "
+                        + "VALUES (?,?,?,?,?,?,?)";
             conectarObtendoId(sql);
             comando.setDouble(1, mov.getValor());           //carregando valor
             Date dataSql = Date.valueOf(mov.getData());     //carregando data
@@ -42,6 +41,7 @@ public class MovimentacaoDaoBD extends DaoBd<Movimentacao> implements Movimentac
             comando.setString(4, mov.getDescricao());       //carregando descricao
             comando.setBoolean(5, mov.isTipo());            //carregando tipo
             comando.setInt(6, mov.getCategoria().getId());  //carregando id da categoria 
+            comando.setInt(7, LoginController.TOKEN_USER_ID);   
             comando.executeUpdate();                        //Adicionando registro ao banco
             ResultSet resultado = comando.getGeneratedKeys(); //Para recuperar o id no banco e adicionar ao objeto
             if (resultado.next()) {
@@ -64,10 +64,11 @@ public class MovimentacaoDaoBD extends DaoBd<Movimentacao> implements Movimentac
     @Override
     public void deletar(Movimentacao mov) {
         try {
-            String sql = "DELETE FROM movimentacao WHERE id = ?";
+            String sql = "DELETE FROM movimentacao WHERE id = ? and user_id = ?";
 
             conectar(sql);
             comando.setInt(1, mov.getId());
+            comando.setInt(2, LoginController.TOKEN_USER_ID);
             comando.executeUpdate();
 
         } catch (SQLException ex) {
@@ -82,7 +83,7 @@ public class MovimentacaoDaoBD extends DaoBd<Movimentacao> implements Movimentac
     public void atualizar(Movimentacao mov) {
         try {
             String sql = "UPDATE movimentacao SET valor=?, data=?, hora=?, descricao=?, tipo=?, idcategoria=? "
-                    + "WHERE id=?";
+                    + "WHERE id=? and user_id=?";
             conectar(sql);
             comando.setDouble(1, mov.getValor());
             Date dataSql = Date.valueOf(mov.getData());
@@ -93,6 +94,7 @@ public class MovimentacaoDaoBD extends DaoBd<Movimentacao> implements Movimentac
             comando.setBoolean(5, mov.isTipo());            //carregando tipo
             comando.setInt(6, mov.getCategoria().getId());  //carregando id da categoria
             comando.setInt(7, mov.getId());
+            comando.setInt(8, LoginController.TOKEN_USER_ID);
             comando.executeUpdate();
         } catch (SQLException ex) {
             System.err.println("Erro de Sistema - Problema ao atualizar paciente no Banco de Dados!");
@@ -105,10 +107,10 @@ public class MovimentacaoDaoBD extends DaoBd<Movimentacao> implements Movimentac
     @Override
     public List<Movimentacao> listar() {
         List<Movimentacao> listaMovimentacoes = new ArrayList<>();
-        String sql = "SELECT * FROM movimentacao";
-        
+        String sql = "SELECT * FROM movimentacao where user_id = ?";
         try {
             conectar(sql);
+            comando.setInt(1, LoginController.TOKEN_USER_ID);
             ResultSet resultado = comando.executeQuery();
             while (resultado.next()) {              //id, valor, data, hora, descricao, tipo, idcategoria
                 int id = resultado.getInt("id");                //pegando id
@@ -138,13 +140,15 @@ public class MovimentacaoDaoBD extends DaoBd<Movimentacao> implements Movimentac
     @Override
     public List<Movimentacao> listarPorTipo(Boolean tipoEscolhido) {
         List<Movimentacao> listaMovimentacoes = new ArrayList<>();
-        String sql = "SELECT * FROM movimentacao where tipo = ?";
+        String sql = "SELECT * FROM movimentacao where tipo = ? and user_id = ?";
         try {
             conectar(sql);
             comando.setBoolean(1, tipoEscolhido);
+            comando.setInt(2, LoginController.TOKEN_USER_ID);
             ResultSet resultado = comando.executeQuery();
             while (resultado.next()) {              //id, valor, data, hora, descricao, tipo, idcategoria
                 int id = resultado.getInt("id");                //pegando id
+                int user_id = resultado.getInt("user_id");                //pegando id
                 double valor = resultado.getDouble("valor");    //pegando valor
                 Date dataSql = resultado.getDate("data");       //pegando data
                 LocalDate data = dataSql.toLocalDate();
@@ -154,7 +158,7 @@ public class MovimentacaoDaoBD extends DaoBd<Movimentacao> implements Movimentac
                 boolean tipo = resultado.getBoolean("tipo");
                 int idCategoria = resultado.getInt("idcategoria");
                 Categoria categoria = new CategoriaDaoBD().procurarPorId(idCategoria);
-                Movimentacao moviment = new Movimentacao(id, valor, data, hora, categoria, descricao, tipo);
+                Movimentacao moviment = new Movimentacao(id, user_id, valor, data, hora, categoria, descricao, tipo);
                 listaMovimentacoes.add(moviment);
                 ordenarPorData(moviment, listaMovimentacoes);
             }
@@ -171,15 +175,17 @@ public class MovimentacaoDaoBD extends DaoBd<Movimentacao> implements Movimentac
 
     @Override
     public Movimentacao procurarPorId(int id) {
-        String sql = "SELECT * FROM movimentacao WHERE id = ?";
+        String sql = "SELECT * FROM movimentacao WHERE id = ? and user_id = ?";
         try {
             conectar(sql);
             comando.setInt(1, id);
+            comando.setInt(2, LoginController.TOKEN_USER_ID);
 
             ResultSet resultado = comando.executeQuery();
 
             if (resultado.next()) {
                 int idReceita = resultado.getInt("id");                //pegando id
+                int user_id = resultado.getInt("user_id");                //pegando id
                 double valor = resultado.getDouble("valor");    //pegando valor
                 Date dataSql = resultado.getDate("data");       //pegando data
                 LocalDate data = dataSql.toLocalDate();
@@ -189,7 +195,7 @@ public class MovimentacaoDaoBD extends DaoBd<Movimentacao> implements Movimentac
                 boolean tipo = resultado.getBoolean("tipo");
                 int idCategoria = resultado.getInt("idcategoria");
                 Categoria categoria = new CategoriaDaoBD().procurarPorId(idCategoria);           
-                Movimentacao moviment = new Movimentacao(idReceita, valor, data, hora, categoria, descricao, tipo);
+                Movimentacao moviment = new Movimentacao(idReceita, user_id, valor, data, hora, categoria, descricao, tipo);
                 return moviment;
             }
         } catch (SQLException ex) {
@@ -200,22 +206,6 @@ public class MovimentacaoDaoBD extends DaoBd<Movimentacao> implements Movimentac
         }
         return (null);
     }
-
-    @Override
-    public Movimentacao procurarPorData(LocalDate data) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    @Override
-    public Movimentacao procurarPorValor(Double valor) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    @Override
-    public List<Movimentacao> getListaMovimentacoes(CategoriaDAO listaCat) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-    
     
     // UTIL    
     private void ordenarPorData(Movimentacao Mov, List<Movimentacao> listaMovimentacoes) {
@@ -251,38 +241,5 @@ public class MovimentacaoDaoBD extends DaoBd<Movimentacao> implements Movimentac
         };
         Collections.sort(listaMovimentacoes, cmp);
     }
-    
-//    @Override
-//    public List<Movimentacao> getListaMovimentacoes(CategoriaDAO listaCat) {
-//        List<Movimentacao> listaMovimentacoes = new ArrayList<>();
-//        CategoriaDaoBD x = new CategoriaDaoBD();      
-//        String sql = "SELECT * FROM movimentacao";
-//        
-//        try {
-//            conectar(sql);
-//            ResultSet resultado = comando.executeQuery();
-//            while (resultado.next()) {
-//                double valor = resultado.getDouble("valor");
-//                LocalDateTime datahora = resultado.getTimestamp("datahora").toLocalDateTime();
-//                String descricao =  resultado.getString("descricao");
-//                int idcat = resultado.getInt("idcategoria");
-//                Categoria cat = x.getCategoria(idcat);
-//                        
-//               
-////                id
-//
-//                Movimentacao mov = new Movimentacao() {};
-//                listaMovimentacoes.add(mov);      
-//            }
-//        } catch (SQLException ex) {
-//            System.err.println("Erro de Sistema - Problema ao buscar os pacientes do Banco de Dados!");
-//            throw new BDException(ex);
-//        } finally {
-//            fecharConexao();
-//        }
-//        return listaMovimentacoes;
-//           throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-//    }
-
-
+   
 }
